fix for codedex issues

+++ services/test/playback_test.cpp
@@ -74,7 +74,7 @@ static int32_t InitPlayback(std::unique_ptr<AudioServiceClient> &client, AudioSt
 int32_t StartPlayback(std::unique_ptr<AudioServiceClient> &client, FILE *wavFile)
 {
     uint8_t* buffer = nullptr;
-    int32_t n = 2;
+    size_t n = 2;
     size_t bytesToWrite = 0;
     size_t bytesWritten = 0;
     size_t minBytes = 4;
